Added non-pointter variable self elevation, so you can now use:
int a = 132
int ptr b = a->address

This will elevate 'a' ptr count to same as the 'b' ptr count, and thus
just moving the address of a to b.

This is only needed to do if you are using Allow_Inconsistancies=False (Defaults to True).

normally just do:

int a = 123
int ptr b = a